third_party/blink/web_tests/external/wpt/css/css-properties-values-api/at-property-cssom-expected.txt

This is a testharness.js-based test.
[FAIL] Rule for --no-descriptors is invalid
  assert_true: expected true got false
[FAIL] Rule for --no-syntax is invalid
  assert_true: expected true got false
[FAIL] Rule for --no-inherits is invalid
  assert_true: expected true got false
[FAIL] Rule for --no-initial-color-value is invalid
  assert_true: expected true got false
[FAIL] Rule for --syntax-only is invalid
  assert_true: expected true got false
[FAIL] Rule for --inherits-only is invalid
  assert_true: expected true got false
[FAIL] Rule for --initial-value-only is invalid
  assert_true: expected true got false
[PASS] Rule for --valid has expected cssText
[PASS] Rule for --valid-reverse has expected cssText
[PASS] Rule for --valid-universal has expected cssText
[PASS] Rule for --valid-whitespace has expected cssText
[PASS] Rule for --vALId has expected cssText
[PASS] Rule for --no-initial-universal-value has expected cssText
[PASS] Rule for --tab	tab has expected cssText
[PASS] CSSRule.type returns 0
[PASS] Rule for --valid returns expected value for CSSPropertyRule.name
[PASS] Rule for --valid-reverse returns expected value for CSSPropertyRule.name
[PASS] Rule for --valid-universal returns expected value for CSSPropertyRule.name
[PASS] Rule for --valid-whitespace returns expected value for CSSPropertyRule.name
[PASS] Rule for --vALId returns expected value for CSSPropertyRule.name
[PASS] Rule for --no-initial-universal-value returns expected value for CSSPropertyRule.name
[PASS] Rule for --valid returns expected value for CSSPropertyRule.syntax
[PASS] Rule for --valid-reverse returns expected value for CSSPropertyRule.syntax
[PASS] Rule for --valid-universal returns expected value for CSSPropertyRule.syntax
[PASS] Rule for --valid-whitespace returns expected value for CSSPropertyRule.syntax
[PASS] Rule for --vALId returns expected value for CSSPropertyRule.syntax
[PASS] Rule for --no-initial-universal-value returns expected value for CSSPropertyRule.syntax
[PASS] Rule for --valid returns expected value for CSSPropertyRule.inherits
[PASS] Rule for --valid-reverse returns expected value for CSSPropertyRule.inherits
[PASS] Rule for --valid-universal returns expected value for CSSPropertyRule.inherits
[PASS] Rule for --valid-whitespace returns expected value for CSSPropertyRule.inherits
[PASS] Rule for --vALId returns expected value for CSSPropertyRule.inherits
[PASS] Rule for --no-initial-universal-value returns expected value for CSSPropertyRule.inherits
[PASS] Rule for --valid returns expected value for CSSPropertyRule.initialValue
[PASS] Rule for --valid-reverse returns expected value for CSSPropertyRule.initialValue
[PASS] Rule for --valid-universal returns expected value for CSSPropertyRule.initialValue
[PASS] Rule for --valid-whitespace returns expected value for CSSPropertyRule.initialValue
[PASS] Rule for --vALId returns expected value for CSSPropertyRule.initialValue
[PASS] Rule for --no-initial-universal-value returns expected value for CSSPropertyRule.initialValue
Harness: the test ran to completion.
